feat: abi and integratorDay fixes
- Merged the abis for the v2.2 and v2 events
- Fixed wrong tracking fuel balances for IntegratorDay entity
- Update startblock for v2.2 on development
