+++ config/polygon-development.json
@@ -23,6 +23,6 @@
     }
   },
   "v2_2": {
-    "startBlock": 7837562
+    "startBlock": 7360255
   }
 }
