+++ subgraph.yaml.mustache
@@ -515,31 +515,11 @@ templates:
           handler: handleEventDataUpdated
         - event: Transfer(indexed address,indexed address,indexed uint256)
           handler: handleTransfer
-
-  - name: EventImplementationV2_2
-    kind: ethereum/contract
-    network: {{ network }}
-    source:
-      abi: EventImplementation
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.5
-      language: wasm/assemblyscript
-      file: ./src/mappings/v2.2/eventImplementation.ts
-      entities:
-        - Protocol
-        - ProtocolDay
-        - Integrator
-        - IntegratorDay
-        - Event
-        - Ticket
-        - UsageEvent
-      abis:
-        - name: EventImplementation
-          file: ./abis/v2.2/EventImplementation.json
-      eventHandlers:
-        - event: PrimarySale((uint256,bytes32,address,uint64,uint40)[],uint256,uint256,uint256,uint256)
-          handler: handlePrimarySale
+          {{! V2_2 Events }}
+        - event: PrimarySale((uint256,bytes32,address,uint64,uint40)[],uint256,uint256,uint256,uint256) 
+          handler: handlePrimarySaleV2_2
         - event: SecondarySale((uint256,bytes32,address,uint64,uint40)[],uint256,uint256,uint256,uint256)
-          handler: handleSecondarySale
+          handler: handleSecondarySaleV2_2
+        - event: Claimed((uint256,bytes32,address,uint64,uint40)[])
+          handler: handleClaimedV2_2
 {{ /v2.EventFactory }}
